Complete yaml parameters to test
Two files :
- parameters.yaml
- variables_parameters.yaml
