+++ openfisca_core/tests/assets/parameters.yaml
@@ -1,0 +1,85 @@
+
+# Regular parameter values : meta + VALUES handling variation of its value in time
+- variable: smic_horaire_brut
+  description: SMIC horaire brut
+  format: float
+  type: monetary
+  VALUES:
+    - {deb: 2016-01-01, fuzzy: true,     valeur: 9.67}
+    - {deb: 2015-01-01, fin: 2015-12-31, valeur: 9.61}
+    - {deb: 2014-01-01, fin: 2014-12-31, valeur: 9.53}
+    - {deb: 2013-01-01, fin: 2013-12-31, valeur: 9.43}
+    # etc.
+
+# BAREME node with 1 bracket...
+- variable: famille
+  description: Allocations familiales
+  BAREME:
+    - SEUIL:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: 0}
+          - {deb: 1992-01-01, fin: 2014-12-31, valeur: 0}
+      TAUX:
+        VALUES:
+          - {deb: 2014-01-01, fuzzy: true, valeur: 0.0525}
+          - {deb: 1993-07-01, fin: 2013-12-31, valeur: 0.054}
+
+
+# ... which is annoying for this regular pattern of "base * rate"
+# TODO implement this !
+
+- variable: famille
+  description: Allocations familiales
+  LIN:
+    ASSIETTE: assiette_cotisations_sociales # or alternatively pass the base to calc()
+    VALUES:
+      - {deb: 2018-01-01, fuzzy: true, valeur: 0.019}
+      - {deb: 2017-01-01, fin: 2017-12-31, valeur: 0.019}
+      - {deb: 2016-01-01, fin: 2016-12-31, valeur: 0.0185}
+
+# LIN with PLAFOND can be introduced too:
+
+- variable: vieillesse_plafonnee
+  description: Cotisation plafonnée de vieillesse
+  LIN:
+    ASSIETTE: assiette_cotisations_sociales
+    PLAFOND:
+      VALUE: 1
+    VALUES:
+      - {deb: 2018-01-01, fuzzy: true, valeur: 0.0855}
+      - {deb: 2017-01-01, fin: 2017-12-31, valeur: 0.085}
+
+
+# of course real BAREMEs with multiple brackets happen :
+
+- variable: agffc
+  description: AGFF cadres
+  BAREME:
+    - SEUIL:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: 0}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: 0}
+      TAUX:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: .016}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: .012}
+
+    - SEUIL:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: 1}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: 1}
+      TAUX:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: .017}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: 0.013}
+
+    - SEUIL:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: 4}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: 4}
+      TAUX:
+        VALUES:
+          - {deb: 2015-01-01, fuzzy: true, valeur: 0}
+          - {deb: 2001-04-01, fin: 2014-12-31, valeur: 0}
+
+# Unfortunately things in this world are more complex... go check variable_parameters.yaml.
